Better Guild Composability

+++ events/messageCreate.js
@@ -18,51 +18,62 @@ const sensitiveChannels = [
     config.channel.sensitive.giveawaysID,
 ]
 
-async function warn(message) {
-    if (message.author.id in naughtyList) {
-        if (naughtyList[message.author.id] > 3) {
-            try {
-                const member = message.guild.members.cache.get(
-                    message.author.id
-                )
-
-                member.timeout(1 * 60 * 60 * 1000, 'You are on a timeout.')
-
-                delete naughtyList[message.author.id]
-            } catch (error) {
-                console.log(error)
-            }
-        } else {
-            naughtyList[message.author.id]++
-        }
-    } else {
-        naughtyList[message.author.id] = 1
-    }
-    message.delete()
-
-    await message.channel.send({
-        embeds: [
-            new MessageEmbed()
-                .setColor(config.embed_color)
-                .setDescription(`You can't send a link in this server.`),
-        ],
-    })
-}
-
-client.on('messageCreate', async (message) => {
+async function scan(message) {
     try {
+        const guild = client.guilds.cache.get(config.guildID)
+
         if (
-            message.author.id != config.receptionBotID &&
-            message.author.id != config.ownerID
+            message.author.id != client.user.id &&
+            message.author.id != guild.ownerId
         ) {
             if (
                 sensitiveChannels.indexOf(message.channelId) !== -1 ||
                 message.content.includes('http') ||
                 message.webhookId
             ) {
-                await warn(message)
+                if (message.author.id in naughtyList) {
+                    if (naughtyList[message.author.id] > 3) {
+                        try {
+                            const member = message.guild.members.cache.get(
+                                message.author.id
+                            )
+
+                            member.timeout(
+                                1 * 60 * 60 * 1000,
+                                'You are on a timeout.'
+                            )
+
+                            delete naughtyList[message.author.id]
+                        } catch (error) {
+                            console.log(error)
+                        }
+                    } else {
+                        naughtyList[message.author.id]++
+                    }
+                } else {
+                    naughtyList[message.author.id] = 1
+                }
+                message.delete()
+
+                await message.channel.send({
+                    embeds: [
+                        new MessageEmbed()
+                            .setColor(config.embed_color)
+                            .setDescription(
+                                `You can't send a link in this server.`
+                            ),
+                    ],
+                })
             }
         }
+    } catch (error) {
+        console.log(error)
+    }
+}
+
+client.on('messageCreate', async (message) => {
+    try {
+        await scan(message)
     } catch (error) {
         console.log(error)
     }
@@ -70,18 +81,7 @@ client.on('messageCreate', async (message) => {
 
 client.on('messageUpdate', async function (oldMessage, newMessage) {
     try {
-        if (
-            newMessage.author.id != config.receptionBotID &&
-            newMessage.author.id != config.ownerID
-        ) {
-            if (
-                sensitiveChannels.indexOf(newMessage.channelId) !== -1 ||
-                newMessage.content.includes('http') ||
-                newMessage.webhookId
-            ) {
-                await warn(newMessage)
-            }
-        }
+        await scan(newMessage)
     } catch (error) {
         console.log(error)
     }
